Add skeleton structure (#9)
* adding notes about "protocol"

* Adding .gitattributes for git lfs

* Adding architecture diagram

* write an initial project description

* Adding initial python project files

* skeleton of services

* Adding skeleton

<rebased>

* expanded skeleton

added some new services and documentation and expanding the skeleton of
the prototype system

* Updated SmartDiskService and FileInfo factory

* added more to service skeletons

* Documenting more of the protocol

this is partially to document my thoughts but also to get me more in the
mood for writing

* Added a storage.types module

this should help with keeping the services module a bit more readable

* added more structure

* creating stub for fs backend

some of the code in services will be migrated here, so that the services
module can mostly just contain flight logic. This will also help to
better mock flight services with mixed storage backends.

* adding a script to represent an fs endpoint

* checkpoint on protocol notes

* cleaning submodule state

* cleaning submodule state

* submodules: adding substrait

* mohair/substrait: generated protobuf wrappers

* mohair/substrait: modified import prefixes

without these changes, the 'mohair' package name is missing from the
imports and there is nothing to indicate they should be local to the
module location.

* buf.gen.yaml: added gen config for buf build

This is a modified version of what substrait provides. Simplified to
just python for now

* poetry.lock: updated

* submodules: added mohair-protocol

this submodule is an extension for substrait

* mohair/skytether: added generated wrappers

* mohair/skytether: updated import

* install-deps.r: adding r script for installs

A simple r script that lists the installed packages and can be used to
install them