mohair/query: minimal substrait functionality
added code to get some minimal amount of substrait functionality.

`operators.py` contains code where we will be able to do transformations
of query plans, and also produce an acero execution plan. Currently, it
only takes a substrait plan and produces a representative internal
structure that maintains references to the protobuf structure.

`planner.py` will be the interface to the query planner and optimizer,
but currently only contains code that translates a substrait plan
(protobuf structure) to a mohair plan (internal structure).

`types.py` and `executor.py` are stubs.