simple query caching and logging
Added naive query caching mechanism to be able to get results in a
separate request (DoGet) instead of the same request (DoAction). The
mechanism hashes operators in a query plan and uses the result hash as
an identifier. This is a super simple approach that can also be used as
a naive approach for reusing materialized results.