task: add classes for processing query plans
Added classes, implemented similarly to visitor pattern, to walk an
operator tree and calculate tree properties. These are then used to view
the query plan or split the query plan.

Splitting a query plan, at the moment, defaults to finding an N-ary
operator (e.g. join or set) with the tallest height. This represents an
operator that would be the leaf of the `SuperPlan` and whose inputs
would each be the root of a `SubPlan`. Each `SubPlan` can be remotely
executed on a single dataset and the query results would be sent
upstream to be gathered in the N-ary operator such that network and data
access latencies can be somewhat hidden.

Issue: #22
Related: #24