task: simplify representation of decomposed plans
To facilitate constructing new substrait plan messages easier, a subplan
should have a reference to the root of the original query plan parsed
from substrait. Because of this, there is no benefit to representing
SuperPlan and SubPlan types separately. An instantiation of a SubPlan
can be viewed as the substrait plan message constructed from a
particular tree in a split.

Issue: #22