feat: uptake new mohair message, `PlanAnchor`
The new message is currently assigned to the "optimization" property of
a substrait Plan. The purpose of the new message is to be able to
propagate the leaf operator of the super plan that pushback plans should
be merged into without complicating semantics.

Also, updated the configuration in buf.gen.yaml