task: refactored clients and services
The refactoring here also had to accommodate how the plan hash (used to
name results) was being propagated and how the new query processing
types (specifically `SubstraitPlan`) would be used. The current state of
services (DBMSService) is to receive a substrait plan and execute it and
provide the hash of the executed plan for subsequent data retrieval.
This should work fine for the short term, but I will eventually have to
integrate how to split a query plan, etc. in this workflow. So, the
handler may be broken up or I may simply add more functions to make the
workflow understandable.

Issue: #1