task: major updates to mohair plan modeling
I was previously trying to use std::variant to represent the types of
input ops a query op could have, which was real silly. This re-designs
to actually use polymorphism but loses out on empty class optimization
(which was something I wanted to try). Some possible optimizations in
code architecture may be lost with this approach but at least it's
functional.