task: added faodel code for engine execution
Added (or updated) convenience functions to:
- publish an arrow::Table to a kelpie::Pool
- register a function to execute a substrait plan with Acero with kelpie
- invoke the registered function

This is not yet functional or tested because I need to still construct
the initial table for publishing (mostly ready) and to be able to
extract the correct key name from the substrait plan (todo).

Once the key name can be extracted from a substrait plan this code can
be tested.