task: starting to add query plan traversal
I thought this was complete but it is not. It is necessary to traverse
the query plan to extract key names (for kelpie pool) from the query
plan, since the key name is required to invoke the execution function
(an unfortunate oversight on my part).

While I'm doing the traversal to grab key names, I might as well also
prepare code for query splitting which I can more or less port from
python code