task: added code for walking query plan
traversal of query plans and structures for referencing operators and
tree properties are done in plans.[ch]pp. We store pointers to
substrait::Plan (in mohair::QueryPlan and derived classes) and to
substrait::Rel (in mohair::QueryOp and derived classes) for now for
convenience and until the implications for memory management are better
understood.