task: update build
This mostly adds changes to the build to make Faodel integration
optional. To do this we add the following files:
- mohair.options
- mohair-config.hpp.in

We also refactor source that uses Foadel to try and uptake these changes
by adding preprocessor macros to use faodel code only if USE_FAODEL is
defined.

This is unfinished, but likely to be integrated with other changes
before continuing