fix: fixed updates to build and refactoring
It has been awhile since I built on macos, so the build needed some
overall fixes, including upgrading arrow. Oddly, building Arrow 12 seems
to be problematic for poetry so upgrading to Arrow 14 helps, though
there are likely some build issues to still work through and things to
update.

The other major thing in this commit is refactoring to accommodate file
name changes and a few file moves (mohair.hpp and util.cpp into the root
cpp src dir). This is mostly in preparation for integrating back into
the mohair repo as well as making the build more flexible and adjusting
the architecture to better accommodate when I start integrating with
duckdb