task: read Arrow IPC stream into duckdb
This reads an IPC stream from a file (just as binary) and then passes
the buffers to duckdb adapter code which knows how to interact with the
arrow extension for duckdb.

In the future, this should be enhanced so that it doesn't *have* to go
through duckdb, or that it only does so with a certain flag, but for now
this was a prototype implementation.