task: modified `FromReadMsg` to not use NamedTable
Instead of the `named_table` property of `ReadRel`, I change substrait
messages to use the `local_files` property so that it is easier to
control access path selection on the duckdb side.

Because duckdb vendors such an odd version of protobuf (3.19.2), I am
unable to generate custom protobuf definitions for mohair extensions.
Custom parsing of `named_table` would be hack-y and be fairly
incompatible with how duckdb currently works with substrait. So, the
best option is to specify skytether partitions as using Arrow files and
then customizing that code path on the duckdb side (which is currently
totally unused). This also has potential value to the community at
large.