task: allow service to register with topo-service
topo-service is a simple metadata service that should accept requests to
register and de-register service locations. This commit adds that
capability and prototypes it in csd-service.

An API for simple mohair clients is added to src/cpp/clients so that
services can connect to other services and send requests by calling the
appropriate function. A MohairClient simply wraps a FlightClient and
implements a higher-level API for mohair requests over FlightClient and
FlightServers.

Some minor re-organization of duckdb and mohair services were necessary
to enable starting a server with a given location URI instead of a
default one (local host; any port). Using a provided URI will be
convenient for bootstrapping a set of mohair services to communicate
with each other.