task: reorganize code
This is mostly a re-organization for better design, though it was
partially in an attempt to address potential memory issues (it did not
fix this).

The re-organization spans many files, but mostly separates some external
dependencies (standard lib, arrow, arrow-flight, substrait) into their
own headers and then tries to make the design for the distributed
service architecture a bit more sensible.

The distributed services are architected such that the TopologyService
is a lineage from a ServerAdapter (internal adapter to insulate from
arrow flight) and EngineService is a separate lineage that is used for
various CSE services (backed by particular engines). Faodel-specific and
Acero-specific engines are not yet updated for the new design (may not
yet be problematic), but an AceroService (alternate to DuckDBService)
should eventually be implemented.