task: changed read-arrow to use familiar CLI
Instead of being a simple program with hardcoded in-order options,
`read-arrow` is being updated to a similar CLI as `csd-service`,
`topo-service`, etc.

The main benefit is that now `read-arrow` will be able to read Arrow
data from a variety of inputs without having to change implementation
(IPC stream, IPC files, etc.).