task: updated includes for relocatable install
include directives were previously relative directories which works when
building locally, but does not work when installed in a system (in a way
that is reasonably isolated). In order to change the include directives,
the build definition needed to define the correct install locations,
include directories (for `-I` options), and the correct dependencies.