task: updated write-substrait and supporting code
Updated mohair-substrait (v0.1.8 -> v0.1.9).

Updated mohair.query.extension so that toolbox/write-substrait will
write a SkyRel message instead of a ReadRel message. Split
mohair.query.relations across mohair.query.types and mohair.util because
the code was mostly a duplication of mohair.query.types anyways.

The motivation for changing from ReadRel to SkyRel is to ensure that
query engines can appropriately handle the distinction of a partition
(using skytether) and a query that explicitly scans particular slices.
This is difficult with ReadRel because it requires specifying
`named_table` or a file name, but both approaches make it difficult to
concisely specify handling of a partition (though either can be used for
slices).