task: updated build to find alternate libs
In order to fix a bug, it was necessary to build a dependency manually
and I wanted to install it in an alternate location. This commit
accommodates this scenario by assuming that `pkg_config_path` is
configured (via `-Dpkg_config_path="..."` or `--pkg-config-path "..."`)
in order to customize where the `dependency` function find libraries.
Meson's `dependency` function uses pkg-config to find libraries, and
pkg_config_path specifies locations to look *first*.

The use of `get_compiler('cpp').find_library` was unreliable (could find
some libraries but not all).