task: updated generated protobuf code and deps
The generated protobuf code needed to be updated to accommodate the new
namespace of protobuf types (specifically, mohair.SkyRel ->
skytether.mohair.SkyRel). This is a cascading effect from a fix for C++.

Additionally, some deps needed to be downgraded:
- ibis-substrait (4.0.1 -> 4.0.0)
- substrait (0.23.0 -> 0.20.0)

ibis-substrait 4.0.1 has a bug and is compatible with substrait<=0.20.0
(which is on substrait protocol version 0.52.0). While I need substrait
>= 0.20.0 (due to change from `JOIN_TYPE_SEMI` to `JOIN_TYPE_LEFT_SEMI`
and other related changes), this commit at least works with the
downgraded versions of ibis-substrait and substrait (for the purposes of
writing substrait and splitting substrait plans).