task: added ibis translations for new rel types
This makes it so that on the plan production side, we can be more
explicit about how we want the consumer to resolve data (SkyRel or
SkyPartitionRel/SkySliceRel). Since the write side is independent of the
read side, this is a way for intent to be communicated from the write
side to the read side.