task: updated write-substrait with new rel type
This particular change is more of a test for single-slice partitions
(one data slice and no metadata slices) by specifying it as a
SkySliceRel with the partition name as the slice_key, rather than as a
SkyRel.

This needs to be parameterized and expanded to make it easier for us to
run various experiments without always hack-ing the code.