task: moved `using` statements into a namespace
This is because a `using` statement to alias a type in another namespace
(e.g. to enable omission of the full namespace) is apparently "bad"
practice (as in rude, not as in bad). While I'm not surprised, I think
now is a good time to at least move them into non-global namespaces.
Their convenience far outweighs the drawbacks in the short term.