task: update to query workflow
This commit does 3 things:
1. Introduce a `MohairTicket` which is an extension of flight::Ticket
   with attributes and functions for our system (currently just an int to
   represent a named query context).
2. Factor out some validation and receiving of the initial response of a
   query request into smaller, more convenient functions.
3. Parse a QueryResult into an arrow::RecordBatchReader for sending data
   between flight services as Arrow data.

(1) is primarily for convenience and an approach to insulating Ticket
implementation. (2) is to help find a way of hiding low-level or verbose
logic for interacting with remote services behind more convenient
high-level APIs. (3) is necessary for figuring out how to flow data
between services in a way that makes it easy to leverage flight and for
us to actually complete remote execution of a query (end-to-end for a
single hop client-server architecture).