task: modularized test workload
I changed the write-substrait implementation to be a cleaner layout to
make it clear which portion is generating the workload query itself.

I also added a directory under toolbox, `toolbox/workloads`, to hold
python modules defining workloads in a way that is easy to understand.

I also expanded `SkyDomain` to hold a map of its partitions so that it
feels more natural to pass it around to access whatever partitions we
may have defined. It acts a bit like a catalog and the idea is that
playing with that interaction will make it easier to design a custom
duckdb catalog for skytether in the future.