task: expansive refactoring
This commit is primarily refactoring to change references to "mohair",
"Mohair", and `skytether` to "skytether", "Skytether", and `skyproto`.

This repo will likely eventually be rolled into the skytether repo, and
the library generated by this project has been changed from
`skytether-mohair` to `skytether` to reflect the change in
responsibilities. Most of what was considered "mohair" has been pushed
into mohair-substrait (which has been refactored from `mohair-substrait`
to `mohair`).

These changes are happening now as I work towards splitting query plans
and merging pushback plans because I need to clean up the code
organization. The reason for this is that I realize that duckdb
shouldn't own most of the cooperative query decomposition logic (though
it can provide support). Instead, this library (skytether) should
provide most of that logic and it should be implemented in the service
because it must receive a substrait plan, split it in substrait,
propagate the substrait subplan, receive and merge a substrait pushback
plan, translate to an engine plan (e.g. duckdb), then execute, then form
a pushback plan for response. Many of those steps fall outside of what
an engine should do including manipulating the substrait plan in various
ways as well as translating the "finalized" engine plan back to a
substrait plan.