task: updated translation and pushback creation
Translation to duckdb needed updating to not rely on output aliases of
the root PlanRel, instead using the output schema of the root operator.

Additionally, creation of the pushback plan needed to accommodate the
output schema.

Finally, added updates to python for translating from ibis to substrait
to accommodate the schema field of the various SkyRel types