task: updated skytether service API
Updated the service API to use different values for flight tickets (name
of view instead of simple ID) as well as added an implementation for the
actual cooperative decomposition of a query plan.

The responsibility of the service API is to (optionally) split a
received query plan (eager split) to propagate the subplan downstream.
Then, merge the pushback plan as necessary before passing a final query
plan to the engine for execution. The engine should respond with a
pushback plan, which the service API is responsible for sending to the
upstream device as a response.