task: added context ID to duckdb-mohair CLI
duckdb-mohair was previously relying on the old SkytetherTicket
(materialization name only). For now, context ID is still required for
a skytether service to retrieve a query context, so duckdb-mohair needed
a way for the context ID to be specified and it needed to update its
SkytetherTicket usage. Additionally, I updated SkytetherTicket with some
convenience code since the old approach is unnecessary (ticket data
consisting of only a stringified context ID).