task: cleaned up some hacky debug code
I removed unnecessary code that made it easier to inspect intermediate
objects in a debugger. While they are convenient for that purpose, I am
more familiar with the debugger now and also the removed code was
cluttering up the source.

Instead, a future commit should re-add the code in a cleaner way that
can easily be turned off (verbose and terse modes for debugging) and
preferably add logging instead of just writing to stdout/stderr.