task: added support for logging perf timestamps
The added code mirrors the implementation in mohair and uses functions
from mohair as well. It's not very clean, but for now it's functional. I
will probably redesign it later (maybe move it out of mohair entirely).
I also want to move some of the performance metrics into runtime
statistics included in pushdown and pushback plans.