task: changed topology config to toml
This changes the format of the topology config to a *simplified* subset
of TOML. I do not currently include a full TOML parser nor implement
one. This change is to make it easier to add new config parameters, such
as human-readable labels for each engine, the location of the topology
metadata server, and a path to a data directory.

Some of these are not yet used, but this will allow me to set up better
experiment infrastructure.