task: updated annotation of manual splits
Previously, all desired split points were annotated upfront, then after
each iteration annotations were removed. This was silly and didn't
properly explore the space of split points.

Now, the number of split points is set to 2 (hard-coded for now) to
match a 3-level topology (N - 1 levels is how many splits we will make).
Then, each iteration annotates exactly the 2 operators at which we want
to make splits. This way we can uniquely name each iteration, we can
clearly validate if the resulting execution plans are correct or not,
and we can fully explore the space of plan splits.